correct path trace

+++ scls/path_trace_render.cl
@@ -138,9 +138,7 @@ __kernel void   path_trace_render(
 			dir = dir - (*screen).center;
 			dir = fast_normalize(dir);
 
-	//		unsigned int seed0 = tx * 3 % WIDTH + (rands.x * (int)(WIDTH * 0.1f));
-			unsigned int seed0 = tx * 3 % WIDTH + (rands.x * WIDTH * 0.1f);
-	//		unsigned int seed1 = ty * 3 % HEIGHT + (rands.y * (int)(HEIGHT * 0.1f));
+			unsigned int seed0 = tx * 3 % WIDTH + (rands.x * (int)(WIDTH * 0.1f));
 			unsigned int seed1 = ty * 3 % HEIGHT + (rands.y * (int)(HEIGHT * 0.1f));
 			for (int k = 0; k <= N; k++)
 			{
